requirements.txt
-f /usr/share/pip-wheels
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
idna==2.8
joblib==0.14.0
mysqlclient==1.4.4
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
SQLAlchemy==1.3.8
urllib3==1.25.6
